Move Expression2SqlSerializer into builder
